+++ test/hirefire/macro/test_delayed_job.rb
@@ -71,16 +71,40 @@ class HireFire::Macro::Delayed::JobTest < Minitest::Test
     assert_equal 0, HireFire::Macro::Delayed::Job.job_queue_size
   end
 
-  def test_job_queue_size_counts_locked_jobs
+  def test_job_queue_size_excludes_locked_jobs
     BasicJob.delay.perform.update(locked_at: Time.now, locked_by: "worker-1")
-    assert_equal 1, HireFire::Macro::Delayed::Job.job_queue_size
+    assert_equal 0, HireFire::Macro::Delayed::Job.job_queue_size
   end
 
-  def test_job_queue_latency_counts_locked_jobs
+  def test_job_queue_latency_excludes_locked_jobs
     Timecop.freeze(1.minute.ago) do
       BasicJob.delay.perform.update(locked_at: Time.now, locked_by: "worker-1")
     end
+    assert_equal 0, HireFire::Macro::Delayed::Job.job_queue_latency
+  end
+
+  def test_job_queue_size_counts_due_unlocked_only_with_locked_and_future_present
+    BasicJob.delay(queue: :default).perform
+    BasicJob.delay(queue: :mailer).perform
+    BasicJob.delay(queue: :default, run_at: 1.minute.from_now).perform
+    BasicJob.delay(queue: :other).perform.update(locked_at: Time.now, locked_by: "worker-1")
+
+    assert_equal 2, HireFire::Macro::Delayed::Job.job_queue_size
+    assert_equal 1, HireFire::Macro::Delayed::Job.job_queue_size(:default)
+    assert_equal 1, HireFire::Macro::Delayed::Job.job_queue_size(:mailer)
+    assert_equal 0, HireFire::Macro::Delayed::Job.job_queue_size(:other)
+    assert_equal 2, HireFire::Macro::Delayed::Job.job_queue_size(:default, :mailer)
+  end
+
+  def test_job_queue_latency_ignores_locked_when_unlocked_due_exists
+    Timecop.freeze(3.minutes.ago) do
+      BasicJob.delay(queue: :default).perform.update(locked_at: Time.now, locked_by: "worker-1")
+    end
+    Timecop.freeze(1.minute.ago) { BasicJob.delay(queue: :mailer).perform }
+
     assert_in_delta 60, HireFire::Macro::Delayed::Job.job_queue_latency, LATENCY_DELTA
+    assert_equal 0, HireFire::Macro::Delayed::Job.job_queue_latency(:default)
+    assert_in_delta 60, HireFire::Macro::Delayed::Job.job_queue_latency(:mailer), LATENCY_DELTA
   end
 
   def test_raises_when_no_mapper_is_detected
@@ -99,6 +123,15 @@ class HireFire::Macro::Delayed::JobTest < Minitest::Test
     elsif defined?(Mongoid)
       assert_equal 1, HireFire::Macro::Delayed::Job.queue(:default, mapper: :mongoid)
     end
+  end
+
+  def test_deprecated_queue_method_excludes_locked_jobs
+    BasicJob.delay(queue: :default).perform
+    BasicJob.delay(queue: :default).perform.update(locked_at: Time.now, locked_by: "worker-1")
+
+    mapper = defined?(ActiveRecord) ? :active_record : :mongoid
+
+    assert_equal 1, HireFire::Macro::Delayed::Job.queue(:default, mapper: mapper)
   end
 
   def test_deprecated_queue_method_with_priority_range